chore: move getUpdatedMessages to helpers
